Fix completions installation condition
Recently zinit started installing completions event if nocompletions
ice is specified. This seems to be related to recent commit
https://github.com/zdharma/zinit/commit/a46f97e8c37c51f63788599a2b2870f6f7c7a39e

With that change, if either nocompletions or as"null" ices are not
specified than completions are installed, so to avoid completions
installation one has to specify both nocompletions _and_ as"null".

This commit changes logic to be equivalent to what was before.