hide internal function arguments from sourced scripts
When manually sourcing a file from `.zshrc`, `$@` is empty during
the evaluation of the file. However, when sourcing the same file
via `zinit snippet`, `zinit light` or `zinit load`, `$@` are those
of the internal `zinit` function that happens to invoke `source`.
This PR sets `$@` to empty before sourcing scripts to make sourcing
of files via `zinit` consistent with manual sourcing.

Test case:

  % cat ~/.zshrc
  source ~/.zinit/bin/zinit.zsh
  source ~/plugin.zsh
  zinit snippet ~/plugin.zsh
  % cat ~/plugin.zsh
  print -lr -- start "$@" end

Without this commit:

  % zsh
  start
  end
  start
  /home/romka/plugin.zsh
  end

With this commit:

  % zsh
  start
  end
  start
  end

Motivation:

Some zsh scripts allow additional arguments to be passed when they
are being sourced.

  source ~/plugin.zsh foo bar

Sourcing such scripts via `zinit` is currently impossible. It would
be perfect if `zinit` allowed custom arguments to be passed to
sourced scripts. This PR doesn't go that far but at least it allows
passing no arguments.