support OSTYPE linux-musl
Typically, when zsh is built as a static binary (as with https://github.com/romkatv/zsh-bin)
OSTYPE reports linux-musl, instead of the more common linux-gnu.
Which in turn breaks binary selection from gh-r, and leads to download of mac binaries.

This change maps both linux-gnu and linux-musl to the match string "linux".

This fixes #416