fix: Don't escape exclamation marks in hook keys
Exclamation marks (per default) get escaped in interactive mode,
but not in non-interactive modes. In this case it meant that zinit
would not find any hooks which implements the ices when running in
non-interactive mode (e.g. in a script). This in turn would NOT result
in errors but would just silently not execute any of the ices which
of course would be a bad thing(tm).

https://github.com/zdharma-continuum/zinit/issues/199 has the details...

Closes: #199