feat: feature – support miligrams – appimage compressed plugins
This is the first commit that introduces the miligrams support via
new, embedded (shipped and loaded by default) `mg` annex. In this commit
what's implemented is:
- loading the plugin from a miligram via normal `zi for my-plugin.mg`
  syntax, with a variation of `zi mg for my-plugin`,
- keeping the miligram mounted via `keep` ice (i.e.: `zi keep for
  my-plugin.mg`),
- a new subcommand `xmg`, that can build the miligram (via `zi xmg
  build` in the plugin dir), and also:
- … that can extract miligram (via `zi xmg extract my-plugin`).

What's left is only allowing download of the miligrams from the GitHub
Releases page. The planned syntax for this is: `zi mg"user/repo" for
my-plugin`, that would look into the binary files and look up the
most recent version of `my-plugin.mg`, like e.g.: `my-plugin-v1.7.mg`.